Remove git diff in update-release job
The git diff command seems to make the script hang indefinitely when the
diff output is too long to fit and requires scrolling to the
bottom. This happened in the credhub compilation build, where the script hung
waiting for user input and the build never finished; we recently added
bbr to the credhub.yml file and the git diff is siginifcatly larger now.