Configure agent blobstores via env instead of CPI
We are trying to ensure that blobstore credentials do not show up on
deployed VMs.  Currently they show up twice in agent configuration: once
from the CPI configuration and once from the agent env configuration.
This change moves CPI configuration to the agent env since the former is
only supposed to be a fallback mechanism and is deprecated.

[#169347469](https://www.pivotaltracker.com/story/show/169347469)