Fix ops files for Postgres 9 and 10
We've previously updated the BOSH Director manifest to use the new
Postgres 13 job. The existing ops files for Postgres 9 and 10 still
assumed that Postgres 10 was the default.

In the case of Postgres 9, this would insert Postgres 9 in addition to
the default, causing monit to fail with the following error:

```
/var/vcap/monit/job/0000_postgres-9.4.monitrc:2: Error: service name conflict, postgres already defined '/var/vcap/sys/run/bpm/postgres-9.4/postgres-9.4.pid
```

In the case of Postgres 10, it was simply doing nothing because the
previous default was Postgres 10.

Signed-off-by: Brian Upton <bupton@vmware.com>