BOSH Director is deployed with Jammy stemcell
Previously, the BOSH Director was deployed with Bionic Beaver-based
Ubuntu stemcells.

This change enables the open source community access to the latest
stemcells. The Jammy-based stemcells are the newest and most up-to-date
line of stemcells.

Although the BOSH Director is deployed on Jammy, this does not prevent
the Director from deploying VMs with Bionic- or Xenial-based stemcells.

An important aspect is ensuring that the CPIs work on a Jammy Director.
We are confident that the existing CPIs, with
the exception of vCloud, will work.

Of the Ruby-based CPIs, we ensured that the Ruby version was >= 3.1,
required by Jammy due to its use of OpenSSL3.

Golang-based CPIs do not link with OpenSSL; Golang uses its own crypto
library. Thus they don't have the same problems that the Ruby-based CPIs
have.

- **AWS** (Ruby) (deployed a VM)
- **Azure** (Ruby) (deployed a VM)
- Docker (Golang)
- GCP (Golang)
- **OpenStack** (Ruby) (checked that Ruby 3.1 was used)
- **vSphere** (Ruby) (deployed a VM)
- warden (Golang)

- alicloud (Golang, but has a .ruby-version of 3.0)
- cloudstack (Golang, https://github.com/orange-cloudfoundry/bosh-cpi-cloudstack)
- SoftLayer (_seems_ to be Golang)
- vCloud (Ruby, not upgraded to 3.1, archived, read-only: https://github.com/cloudfoundry-attic/bosh-vcloud-cpi-release/tree/master/src/bosh_vcloud_cpi) **this will be broken**
- Virtualbox (Golang)

We removed Jammy-specific manifest operations files; they are no longer
needed.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>