Enable users to deploy on Bionic stemcells
We give the users the option to deploy the BOSH Director to the older
Bionic stemcells instead of the newer Jammy stemcells.

We provide instructions in the README for deploying with Bionic.

We provide manifest operations files (`use-bionic.yml`) for all IaaSes
with the exception of SoftLayer, for SoftLayer does not use Bionic but
instead uses the venerable Xenial stemcell (and an old line of the
Xenial stemcell at that).

We updated `vsphere/cpi.yml` to use a jammy stemcell to make testing
easier.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>