Unpin BOSH Docker CPI version
The version of the BOSH Docker CPI had been pinned to 0.0.5 for the last
several years, and artifact of a time when _all_ versions were pinned.
This was a dark time for BOSH Deployment.

The version was moot as long as the repository pointed to was the
dormant `cppforlife/bosh-docker-cpi-release`, whose final release was
0.0.5; however, when we updated to the correct repository whose latest
release was 0.0.7, the improper pinning no longer served us.

This commit unpins the Docker CPI so it can fly free.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>