Remove PowerDNS
The PowerDNS package has not been updated since 2016. This commit removes:
- `misc/powerdns.yml`

This change is related to, but not dependent on, the removal of all
PowerDNS packages, jobs, and cofig from the BOSH repo in this [PR 2501](https://github.com/cloudfoundry/bosh/pull/2501)

The removal of this ops-file is a way to signal that PowerDNS is no
longer supported, and will be removed, see BOSH PR above.